Initial flow for creating a new state
Add a 'UI State' variable which tracks the current state of the overall
UI, e.g., 'normal', 'create region', 'create timespan', etc.

Adds a flow where if you click 'create region', the lower two panes
are greyed out, the viz pane remains the same, and the state pane
is replaced by a place to enter the name of the new region.