Preliminary create region state.
Adds some preliminary code to create a new region state. Allows the
user to click on the visualization to create a fixed-size circular
region. The logic is encapsulated in a CreateRegionInteraction class.

Pulled out some commonly needed info from the 'drawToSVG' function
so that it would be available to both 'drawToSVG' and the new
'CreateRegionInteraction' class.

This is mostly just to get some scaffolding down - I think a lot of
this will need to be reworked :) Ideally, it would be nice to
put the SVG elements representing a region into an actual React
Component, instead of managing them directly (but this is not
so simple because the SVG is managed by d3...)