Allow drag/rotate/translate for data.
Allow the user to edit the data in the visualization.
Sort of overloading the EditBox class to double as a transform.
The initial Lat/Long->SVG coordinates transform is now
handled by an EditBox.

NOTE: This change breaks the region state calculations :)
Next step is to revamp region states using EditBoxes.