Enable move/resize/rotate for regions.
Once a region is created, the user can move/resize/rotate it.
This also fixes the region state calculation.
Removes a decent amount of old code and simplifies viz-view.js.

The flow for creating a region is still a little awkward,
because you have to click on the map first.
It'd be better to drag the mouse to create an initial shape.